+++ CHANGELOG.md
@@ -5,6 +5,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added new internal Write-Log function
+- Added Copy-Object function to copy hashtable objects by value instead of by reference
+- Added Import-PSDataFile function that fixes the limitation of the default Import-PowerShellDataFile function.
+
+### Remove
+
+- Rollback of the new DataObject functions. These will be replaced by the [ObjectGraphTools module](https://www.powershellgallery.com/packages/ObjectGraphTools)
+
 ## [0.2.0] - 2024-01-02
 
 ### Added